requirements.txt
discord.py==0.16.12
giphypop==0.3
requests===2.19.1